refactor(pip): decouple pip installer from config package
This commit refactors the pip installation structure to adhere strictly to the packaging design guideline outlined in doc/howwewritego.md#internalconfig:

"The internal/config package contains only struct types and constants mirroring the YAML schema. Do not add functions or methods to this package."

It shifts all pip installation helpers and table-driven test suites from internal/config into a dedicated utility package internal/pip (as pip.go and pip_test.go). This cleanly decouples execution logic from configurations, keeps code DRY and compliant, and preserves 100.0% package-level statement coverage.
