+++ internal/librarian/java/install.go
@@ -21,6 +21,7 @@ import (
 	"os/exec"
 
 	"github.com/googleapis/librarian/internal/config"
+	"github.com/googleapis/librarian/internal/pip"
 	"github.com/googleapis/librarian/internal/yaml"
 )
 
@@ -39,7 +40,7 @@ func Install(ctx context.Context) error {
 		return fmt.Errorf("parsing embedded librarian.yaml: %w", err)
 	}
 	if len(cfg.Tools.Pip) > 0 {
-		if err := config.InstallPipTools(ctx, cfg.Tools.Pip); err != nil {
+		if err := pip.Install(ctx, cfg.Tools.Pip); err != nil {
 			return fmt.Errorf("failed to install pip tools: %w", err)
 		}
 	}
