+++ internal/pip/pip_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package config
+package pip
 
 import (
 	"errors"
@@ -23,9 +23,10 @@ import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
+	"github.com/googleapis/librarian/internal/config"
 )
 
-func TestInstallPipTools(t *testing.T) {
+func TestInstall(t *testing.T) {
 	tmpDir := t.TempDir()
 	stubLogPath := filepath.Join(tmpDir, "pip_invocations.log")
 	stubContent := fmt.Sprintf(`#!/bin/bash
@@ -42,12 +43,12 @@ echo "pip $@" >> %q
 	t.Setenv("PATH", stubDir)
 	for _, test := range []struct {
 		name     string
-		tools    []*PipTool
+		tools    []*config.PipTool
 		wantArgs string
 	}{
 		{
 			name: "install external packages",
-			tools: []*PipTool{
+			tools: []*config.PipTool{
 				{Name: "PyYAML", Version: "6.0.2"},
 				{Name: "jinja2", Version: "3.1.6"},
 			},
@@ -55,14 +56,14 @@ echo "pip $@" >> %q
 		},
 		{
 			name: "install external packages with raw package spec",
-			tools: []*PipTool{
+			tools: []*config.PipTool{
 				{Name: "synthtool", Package: "git+https://github.com/..."},
 			},
 			wantArgs: "install git+https://github.com/...",
 		},
 		{
 			name: "install package with name only (no version/package)",
-			tools: []*PipTool{
+			tools: []*config.PipTool{
 				{Name: "requests"},
 			},
 			wantArgs: "install requests",
@@ -70,7 +71,7 @@ echo "pip $@" >> %q
 	} {
 		t.Run(test.name, func(t *testing.T) {
 			_ = os.Remove(stubLogPath)
-			err := InstallPipTools(t.Context(), test.tools)
+			err := Install(t.Context(), test.tools)
 			if err != nil {
 				t.Fatal(err)
 			}
@@ -87,7 +88,7 @@ echo "pip $@" >> %q
 	}
 }
 
-func TestInstallPipTools_Error(t *testing.T) {
+func TestInstall_Error(t *testing.T) {
 	tmpDir := t.TempDir()
 	stubDir := filepath.Join(tmpDir, "bin")
 	if err := os.MkdirAll(stubDir, 0755); err != nil {
@@ -99,28 +100,28 @@ func TestInstallPipTools_Error(t *testing.T) {
 	}
 	for _, test := range []struct {
 		name    string
-		tools   []*PipTool
+		tools   []*config.PipTool
 		setup   func(t *testing.T)
 		wantErr error
 	}{
 		{
 			name: "pip command fails",
-			tools: []*PipTool{
+			tools: []*config.PipTool{
 				{Name: "failpkg"},
 			},
 			setup: func(t *testing.T) {
 				t.Setenv("PATH", stubDir)
 			},
-			wantErr: ErrPipInstall,
+			wantErr: ErrInstall,
 		},
 	} {
 		t.Run(test.name, func(t *testing.T) {
 			if test.setup != nil {
 				test.setup(t)
 			}
-			err := InstallPipTools(t.Context(), test.tools)
+			err := Install(t.Context(), test.tools)
 			if !errors.Is(err, test.wantErr) {
-				t.Errorf("InstallPipTools() error = %v, wantErr = %v", err, test.wantErr)
+				t.Errorf("Install() error = %v, wantErr = %v", err, test.wantErr)
 			}
 		})
 	}
